reduced the file size of dashboard modals
