+++ src/apps/dashboard/modal/ready-for-loan-modal/ready-to-loan-modal.tsx
@@ -1,15 +1,70 @@
 import * as React from "react";
-import { FC, ReactNode } from "react";
-import Modal from "../../../../core/utils/modal";
+import { FC } from "react";
+import { useDeleteReservations } from "../../../../core/fbs/fbs";
+import { useDeleteV1UserReservationsIdentifier } from "../../../../core/publizon/publizon";
+import Modal, { useModalButtonHandler } from "../../../../core/utils/modal";
 import { useText } from "../../../../core/utils/text";
+import { getReservationType } from "../../util/helpers";
+import ReadyToLoanModalContent from "./ready-to-loan-modal-content";
 
 export interface ReadyToLoanModalProps {
   modalId?: string;
-  children: ReactNode;
 }
 
-const ReadyToLoanModal: FC<ReadyToLoanModalProps> = ({ modalId, children }) => {
+const ReadyToLoanModal: FC<ReadyToLoanModalProps> = ({ modalId = "" }) => {
   const t = useText();
+  const { close } = useModalButtonHandler();
+  const { mutate: deletePhysicalReservation } = useDeleteReservations();
+  const { mutate: deleteDigitalReservation } =
+    useDeleteV1UserReservationsIdentifier();
+
+  const removeSelectedReservations = (
+    selectedReservations: {
+      [key: string]: string;
+    }[]
+  ) => {
+    const selectedReservationsKeys = Object.keys(selectedReservations);
+    const selectedReservationsValues = Object.values(selectedReservations);
+    if (selectedReservationsKeys.length > 0) {
+      selectedReservationsKeys.map((reservation) => {
+        const index = selectedReservationsKeys.indexOf(reservation);
+        const reservationToDelete = selectedReservationsValues[index];
+        const reservationType = getReservationType(reservation);
+        switch (reservationType) {
+          case "physical":
+            deletePhysicalReservation(
+              {
+                params: { reservationid: [Number(reservationToDelete)] }
+              },
+              {
+                // todo error handling, missing in figma
+                onError: () => {
+                  close(modalId);
+                }
+              }
+            );
+            break;
+          case "digital":
+            deleteDigitalReservation(
+              {
+                identifier: String(selectedReservationsValues)
+              },
+              {
+                // todo error handling, missing in figma
+                onError: () => {
+                  close(modalId);
+                }
+              }
+            );
+            break;
+          default:
+            return false;
+        }
+        close(modalId);
+        return false;
+      });
+    }
+  };
 
   if (!modalId) {
     return null;
@@ -23,7 +78,9 @@ const ReadyToLoanModal: FC<ReadyToLoanModalProps> = ({ modalId, children }) => {
         "materialDetailsModalAriaDescriptionText"
       )}
     >
-      {children}
+      <ReadyToLoanModalContent
+        removeSelectedReservations={removeSelectedReservations}
+      />
     </Modal>
   );
 };
